# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Log Monitoring Platform API
  version: 1.0.0
  description: >-
    REST API contract for the MVP, synchronized with docs/contracts/rest-api.md in Pass 2. Every
    JSON response uses the shared ApiResponse<T> envelope; named *Response schemas describe data
    payloads only. Inline specializations must not generate separate Java envelope classes.
    Services/modules do not return HTTP envelopes or response DTOs. Protected human APIs re-evaluate
    user status (disabled user: 403); disabled login/refresh or ingestion scope returns 401. Logout
    revokes the current refresh session and a supplied valid unexpired access token for the same user.
    Other sessions and historical reads remain authorized. Input length violations return 400 with field-to-message validation data before
    persistence. SSE retains its separate contract; other authoritative docs await Pass 3
    synchronization.
servers:
- url: /api/v1
security:
- bearerAuth: []
...
paths:
  /auth/login:
    post:
      tags:
      - Auth
      security: []
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    success:
                      const: true
                    data:
                      $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
      description: >-
        Authenticate an ACTIVE ADMIN or ENGINEER and issue access and refresh JWTs. Initial ADMIN is created by bootstrap, not REST.
        Invalid credentials or a DISABLED user return 401. Credentials/tokens must never be logged.
  /auth/refresh:
    post:
      tags:
      - Auth
      summary: Refresh authentication tokens
      description: >-
        Authenticate the refresh_token in the JSON body; no Bearer access token is required.
        Successful refresh consumes the current single-use token and returns a rotated pair for the
        same session. The old token cannot be reused, including concurrently. Invalid, expired,
        revoked, consumed tokens or a DISABLED user return 401. Reuse of a consumed refresh JWT
        revokes its refresh session, including the current rotated token; other sessions/devices
        are unaffected. Refresh rotation or replay handling alone does not revoke access tokens.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Refresh authentication tokens successful
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    success:
                      const: true
                    data:
                      $ref: '#/components/schemas/RefreshTokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/logout:
    post:
      tags:
      - Auth
      summary: Logout current session
      description: >-
        Authenticate the current refresh JWT in the JSON body. An optional Authorization: Bearer
        header may carry the current access token; no access token is required. Revoke this refresh
        session, preventing future refresh; other sessions/devices remain active. If the supplied
        access token verifies, is unexpired and belongs to the same user, it cannot be reused after
        successful logout. A valid access token for a different user returns 401 without revocation.
        Missing, invalid or expired optional access tokens do not block logout with a valid refresh
        token. Other access tokens remain subject to expiry and normal authorization checks.
        The client discards local tokens after success. No forced SSE disconnection or revocation
        broadcast is introduced. Invalid, expired, revoked or consumed refresh tokens return 401.
        A required revocation dependency failure returns a safe 500, never a successful logout.
      security:
      - bearerAuth: []
      - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '200':
          description: Logout current session successful
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    success:
                      const: true
                    data:
                      type: 'null'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /applications/{applicationId}/environments/{environmentId}/api-keys:
    get:
      tags:
      - API Keys
      summary: List ingestion API-key metadata
      description: ADMIN only.
      parameters:
      - $ref: '#/components/parameters/ApplicationId'
      - $ref: '#/components/parameters/EnvironmentId'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Credential list
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    success:
                      const: true
                    data:
                      $ref: '#/components/schemas/ApiKeyListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
      - API Keys
      summary: Create ingestion API key
      description: ADMIN only. Requires an active application/environment scope; disabled scope returns 403.
      parameters:
      - $ref: '#/components/parameters/ApplicationId'
      - $ref: '#/components/parameters/EnvironmentId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApiKeyRequest'
      responses:
        '201':
          description: Created; plaintext key returned once
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    success:
                      const: true
                    data:
                      $ref: '#/components/schemas/ApiKeySecretResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api-keys/{credentialId}/rotate:
    post:
      tags:
      - API Keys
      summary: Rotate ingestion API key
      description: ADMIN only. Requires an active application/environment scope; disabled scope returns 403.
      parameters:
      - $ref: '#/components/parameters/CredentialId'
      responses:
        '200':
          description: Rotated; plaintext replacement key returned once
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    success:
                      const: true
                    data:
                      $ref: '#/components/schemas/ApiKeySecretResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api-keys/{credentialId}:
    delete:
      tags:
      - API Keys
      summary: Revoke ingestion API key
      description: ADMIN only.
      parameters:
      - $ref: '#/components/parameters/CredentialId'
      responses:
        '200':
          description: Revoked
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    success:
                      const: true
                    data:
                      $ref: '#/components/schemas/ApiKeyResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
...
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    ApplicationId:
      name: applicationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    EnvironmentId:
      name: environmentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    CredentialId:
      name: credentialId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiResponse'
            - type: object
              properties:
                success:
                  const: false
                data:
                  type: [object, 'null']
                  additionalProperties:
                    type: string
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiResponse'
            - type: object
              properties:
                success:
                  const: false
                data:
                  type: 'null'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiResponse'
            - type: object
              properties:
                success:
                  const: false
                data:
                  type: 'null'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiResponse'
            - type: object
              properties:
                success:
                  const: false
                data:
                  type: 'null'
...
    InternalServerError:
      description: Unexpected server error; no stack traces or secrets are exposed.
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiResponse'
            - type: object
              properties:
                success:
                  const: false
                data:
                  type: 'null'
  schemas:
    ApiResponse:
      type: object
      description: >-
        The single shared HTTP/API envelope ApiResponse<T>. Controllers specialize data with a
        payload DTO; services/modules do not return this type. Inline allOf specializations describe
        wire types and must not generate separate Java envelope classes. HTTP payload mapping may
        use MapStruct when appropriate. SSE events follow their separate contract.
      additionalProperties: false
      required:
      - success
      - message
      - data
      properties:
        success:
          type: boolean
        message:
          type: string
          description: Safe for client display/logging.
        data: {}
    LoginRequest:
      type: object
      required:
      - username
      - password
      properties:
        username:
          type: string
          maxLength: 100
        password:
          type: string
          format: password
    UserSummary:
      type: object
      required:
      - id
      - username
      - role
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        role:
          type: string
          enum:
          - ADMIN
          - ENGINEER
    LoginResponse:
      type: object
      required:
      - access_token
      - refresh_token
      - token_type
      - expires_in
      - user
      properties:
        access_token:
          type: string
          description: Short-lived access JWT; a valid unexpired token supplied at logout for the same user is revoked.
        token_type:
          type: string
          enum:
          - Bearer
        expires_in:
          type: integer
          description: Access-token TTL in seconds; exact TTL is configured during implementation.
        user:
          $ref: '#/components/schemas/UserSummary'
        refresh_token:
          type: string
          description: Current single-use refresh JWT; rotated after each successful refresh. Reuse revokes this refresh session.
...
    ApiKeyResponse:
      type: object
      required:
      - id
      - environment_id
      - key_prefix
      - status
      - created_at
      properties:
        id:
          type: string
          format: uuid
        environment_id:
          type: string
          format: uuid
        key_prefix:
          type: string
        status:
          type: string
          enum:
          - ACTIVE
          - REVOKED
          - EXPIRED
        expires_at:
          type:
          - string
          - 'null'
          format: date-time
        created_at:
          type: string
          format: date-time
        revoked_at:
          type:
          - string
          - 'null'
          format: date-time
    CreateApiKeyRequest:
      type: object
      properties:
        expires_at:
          type:
          - string
          - 'null'
          format: date-time
    ApiKeySecretResponse:
      type: object
      required:
      - credential
      - api_key
      properties:
        credential:
          $ref: '#/components/schemas/ApiKeyResponse'
        api_key:
          type: string
          description: Returned only on create/rotate.
    ApiKeyListResponse:
      type: object
      required:
      - items
      - next_cursor
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ApiKeyResponse'
        next_cursor:
          type:
          - string
          - 'null'
...
    RefreshTokenRequest:
      type: object
      required:
      - refresh_token
      properties:
        refresh_token:
          type: string
          description: Current refresh JWT for this session, supplied in the JSON body.
    LogoutRequest:
      type: object
      required:
      - refresh_token
      properties:
        refresh_token:
          type: string
          description: Current refresh JWT for this session, supplied in the JSON body.
    RefreshTokenResponse:
      type: object
      required:
      - access_token
      - refresh_token
      - token_type
      - expires_in
      properties:
        access_token:
          type: string
          description: Short-lived access JWT; a valid unexpired token supplied at logout for the same user is revoked.
        refresh_token:
          type: string
          description: Current single-use refresh JWT; rotated after each successful refresh. Reuse revokes this refresh session.
        token_type:
          type: string
          enum:
          - Bearer
        expires_in:
          type: integer
          description: Access-token TTL in seconds; exact TTL is configured during implementation.
...